+++ eras/allegra/impl/testlib/Test/Cardano/Ledger/Allegra/ImpTest.hs
@@ -32,6 +32,7 @@ import qualified Data.Map.Strict as Map
 import Data.Sequence.Strict (StrictSeq (..))
 import qualified Data.Set as Set
 import Lens.Micro ((^.))
+import Test.Cardano.Ledger.Allegra.Binary.Annotator ()
 import Test.Cardano.Ledger.Allegra.Era ()
 import Test.Cardano.Ledger.Allegra.TreeDiff ()
 import Test.Cardano.Ledger.Core.KeyPair (KeyPair)