+++ eras/conway/impl/src/Cardano/Ledger/Conway/Scripts.hs
@@ -55,6 +55,7 @@ import Cardano.Ledger.Shelley.Scripts (ShelleyEraScript (..))
 import Cardano.Ledger.TxIn (TxIn)
 import Control.DeepSeq (NFData (..), rwhnf)
 import Data.Aeson (ToJSON (..), (.=))
+import Data.Coerce (coerce)
 import Data.MemPack
 import Data.Typeable
 import Data.Word (Word16, Word32, Word8)
@@ -75,17 +76,17 @@ instance EraScript ConwayEra where
   type NativeScript ConwayEra = Timelock ConwayEra
 
   upgradeScript = \case
-    TimelockScript ts -> TimelockScript $ translateTimelock ts
+    NativeScript ts -> NativeScript $ coerce ts
     PlutusScript (BabbagePlutusV1 ps) -> PlutusScript $ ConwayPlutusV1 ps
     PlutusScript (BabbagePlutusV2 ps) -> PlutusScript $ ConwayPlutusV2 ps
 
   scriptPrefixTag = alonzoScriptPrefixTag
 
   getNativeScript = \case
-    TimelockScript ts -> Just ts
+    NativeScript ts -> Just ts
     _ -> Nothing
 
-  fromNativeScript = TimelockScript
+  fromNativeScript = NativeScript
 
 instance AlonzoEraScript ConwayEra where
   data PlutusScript ConwayEra