+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Era.hs
@@ -63,7 +63,6 @@ module Cardano.Ledger.Api.Era (
 ) where
 
 import Cardano.Ledger.Allegra (AllegraEra)
-import Cardano.Ledger.Allegra.Scripts (translateTimelock)
 import Cardano.Ledger.Allegra.TxAuxData (AllegraTxAuxData (..))
 import Cardano.Ledger.Allegra.TxBody (AllegraEraTxBody (..), ValidityInterval (..))
 import qualified Cardano.Ledger.Allegra.TxBody as Allegra (TxBody (..))
@@ -103,7 +102,7 @@ import Cardano.Ledger.Slot (SlotNo)
 import Control.Arrow (left)
 import Control.Monad (unless, when)
 import Data.Bifunctor (Bifunctor (first))
-import Data.Coerce (coerce)
+import Data.Coerce (Coercible, coerce)
 import Data.Default (def)
 import Data.Kind (Type)
 import qualified Data.Map.Strict as Map
@@ -369,7 +368,7 @@ instance EraApi AlonzoEra where
     mkMemoizedEra @AlonzoEra $
       AlonzoTxAuxDataRaw
         { atadrMetadata = md
-        , atadrTimelock = translateTimelock <$> scripts
+        , atadrNativeScripts = coerce <$> scripts
         , atadrPlutus = mempty
         }
 
@@ -400,15 +399,19 @@ upgradeTxDats ::
 upgradeTxDats (TxDats datMap) = TxDats $ fmap upgradeData datMap
 
 translateAlonzoTxAuxData ::
-  (AlonzoEraScript era1, AlonzoEraScript era2) =>
+  ( AlonzoEraScript era1
+  , AlonzoEraScript era2
+  , Coercible (NativeScript era1) (NativeScript era2)
+  ) =>
   AlonzoTxAuxData era1 ->
   AlonzoTxAuxData era2
-translateAlonzoTxAuxData AlonzoTxAuxData {atadMetadata, atadTimelock, atadPlutus} =
+translateAlonzoTxAuxData AlonzoTxAuxData {atadMetadata, atadNativeScripts, atadPlutus} =
   AlonzoTxAuxData
     { atadMetadata = atadMetadata
-    , atadTimelock = translateTimelock <$> atadTimelock
+    , atadNativeScripts = coerce atadNativeScripts
     , atadPlutus = atadPlutus
     }
+{-# DEPRECATED translateAlonzoTxAuxData "Use upgradeTxAuxData instead" #-}
 
 newtype BabbageTxUpgradeError
   = BTUEBodyUpgradeError BabbageTxBodyUpgradeError